+++ test/EmptyApp20.FunctionalTests/Startup.cs
@@ -20,7 +20,7 @@
             services.AddSingleton<EndpointAddress>(endpointAddress);
 
             var builder = new ConfigurationBuilder();
-            builder.AddApplicationInsightsSettings(instrumentationKey: "Foo", endpointAddress: endpointAddress.ConnectionString, developerMode: true);
+            builder.AddApplicationInsightsSettings(instrumentationKey: InProcessServer.IKey, endpointAddress: endpointAddress.ConnectionString, developerMode: true);
             services.AddApplicationInsightsTelemetry(builder.Build());
         }